ai-pulse/backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.6
anthropic>=0.34.2
httpx>=0.27.2
python-dotenv>=1.0.1
pydantic>=2.10.0
